# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CraKa OSINT Portal API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface OsintApi {
  id: number;
  slug: string;
  name: string;
  url: string;
  command: string;
  example: string;
  pattern?: string | null;
  category: string;
  credits: number;
  isActive: boolean;
  createdAt: string;
}

export interface LookupRequest {
  slug: string;
  query: string;
}

export type LookupResultData = { [key: string]: unknown };

export interface LookupResult {
  data: LookupResultData;
  cached: boolean;
  apiName: string;
  success: boolean;
  developer?: string;
  error?: string | null;
}

export interface HistoryEntry {
  id: number;
  slug: string;
  apiName: string;
  queryVal: string;
  success: boolean;
  createdAt: string;
}

export interface HistoryResponse {
  entries: HistoryEntry[];
  total: number;
  page: number;
  limit: number;
}

export type StatsResponseCategoryBreakdownItem = {
  category: string;
  count: number;
};

export type StatsResponseTopApisItem = {
  apiName: string;
  count: number;
};

export interface StatsResponse {
  totalLookups: number;
  successfulLookups: number;
  activeApis: number;
  cachedResults: number;
  categoryBreakdown: StatsResponseCategoryBreakdownItem[];
  topApis: StatsResponseTopApisItem[];
  recentActivity: HistoryEntry[];
}

export interface AdminLoginRequest {
  username: string;
  password: string;
}

export interface AdminLoginResponse {
  success: boolean;
  token?: string;
  message?: string;
}

export interface CreateApiRequest {
  slug: string;
  name: string;
  url: string;
  command: string;
  example: string;
  pattern?: string | null;
  category: string;
  credits?: number;
  isActive?: boolean;
}

export interface UpdateApiRequest {
  name?: string;
  url?: string;
  command?: string;
  example?: string;
  pattern?: string | null;
  category?: string;
  credits?: number;
  isActive?: boolean;
}

export type AdminStatsResponseCategoryBreakdownItem = {
  category: string;
  count: number;
};

export type AdminStatsResponseTopApisItem = {
  apiName: string;
  count: number;
};

export interface AdminStatsResponse {
  totalLookups: number;
  successfulLookups: number;
  failedLookups: number;
  activeApis: number;
  totalApis: number;
  cachedResults: number;
  totalUsers: number;
  premiumUsers: number;
  categoryBreakdown: AdminStatsResponseCategoryBreakdownItem[];
  topApis: AdminStatsResponseTopApisItem[];
  recentActivity: HistoryEntry[];
}

export interface GrantPremiumRequest {
  referralCode: string;
  plan: string;
}

export type UserDataRecentReferralsItem = {
  date?: string;
  credits?: number;
};

export interface UserData {
  referralCode: string;
  isPremium: boolean;
  premiumPlan?: string | null;
  creditsEarned: number;
  totalReferrals: number;
  referredBy?: string | null;
  recentReferrals?: UserDataRecentReferralsItem[];
}

export interface AdminUserEntry {
  referralCode: string;
  isPremium: boolean;
  premiumPlan?: string | null;
  totalReferrals: number;
  creditsEarned: number;
  createdAt: string;
}

export interface UsersResponse {
  users: AdminUserEntry[];
  total: number;
  page: number;
  limit: number;
}

export interface InitUserRequest {
  sessionId: string;
  referralCode?: string | null;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message: string;
}

export type GetHistoryParams = {
  limit?: number;
  page?: number;
};

export type AdminGetHistoryParams = {
  limit?: number;
  page?: number;
};

export type AdminGetUsersParams = {
  limit?: number;
  page?: number;
};

export type GetUserMeParams = {
  sessionId: string;
};