# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CraKa OSINT Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginRequest"
      responses:
        "200":
          description: Login result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoginResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/apis:
    get:
      operationId: adminListApis
      tags: [admin]
      summary: Admin - list all APIs including inactive
      responses:
        "200":
          description: All APIs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OsintApi"
    post:
      operationId: adminCreateApi
      tags: [admin]
      summary: Admin - create a new API
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApiRequest"
      responses:
        "201":
          description: Created API
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OsintApi"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/apis/{slug}:
    put:
      operationId: adminUpdateApi
      tags: [admin]
      summary: Admin - update an API
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateApiRequest"
      responses:
        "200":
          description: Updated API
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OsintApi"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: adminDeleteApi
      tags: [admin]
      summary: Admin - delete an API
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/history:
    get:
      operationId: adminGetHistory
      tags: [admin]
      summary: Admin - get full history
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: History
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HistoryResponse"
...
  /admin/grant-premium:
    post:
      operationId: adminGrantPremium
      tags: [admin]
      summary: Admin - grant premium to a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GrantPremiumRequest"
      responses:
        "200":
          description: Premium granted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/users:
    get:
      operationId: adminGetUsers
      tags: [admin]
      summary: Admin - list all users
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsersResponse"
...
components:
  schemas:
...
    OsintApi:
      type: object
      properties:
        id:
          type: integer
        slug:
          type: string
        name:
          type: string
        url:
          type: string
        command:
          type: string
        example:
          type: string
        pattern:
          type: string
          nullable: true
        category:
          type: string
        credits:
          type: integer
        isActive:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - slug
        - name
        - url
        - command
        - example
        - category
        - credits
        - isActive
        - createdAt
...
    HistoryEntry:
      type: object
      properties:
        id:
          type: integer
        slug:
          type: string
        apiName:
          type: string
        queryVal:
          type: string
        success:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - slug
        - apiName
        - queryVal
        - success
        - createdAt
    HistoryResponse:
      type: object
      properties:
        entries:
          type: array
          items:
            $ref: "#/components/schemas/HistoryEntry"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - entries
        - total
        - page
        - limit
...
    AdminLoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    AdminLoginResponse:
      type: object
      properties:
        success:
          type: boolean
        token:
          type: string
        message:
          type: string
      required:
        - success
    CreateApiRequest:
      type: object
      properties:
        slug:
          type: string
        name:
          type: string
        url:
          type: string
        command:
          type: string
        example:
          type: string
        pattern:
          type: string
          nullable: true
        category:
          type: string
        credits:
          type: integer
        isActive:
          type: boolean
      required:
        - slug
        - name
        - url
        - command
        - example
        - category
    UpdateApiRequest:
      type: object
      properties:
        name:
          type: string
        url:
          type: string
        command:
          type: string
        example:
          type: string
        pattern:
          type: string
          nullable: true
        category:
          type: string
        credits:
          type: integer
        isActive:
          type: boolean
...
    GrantPremiumRequest:
      type: object
      properties:
        referralCode:
          type: string
        plan:
          type: string
      required:
        - referralCode
        - plan
...
    UsersResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/AdminUserEntry"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - users
        - total
        - page
        - limit
    AdminUserEntry:
      type: object
      properties:
        referralCode:
          type: string
        isPremium:
          type: boolean
        premiumPlan:
          type: string
          nullable: true
        totalReferrals:
          type: integer
        creditsEarned:
          type: integer
        createdAt:
          type: string
      required:
        - referralCode
        - isPremium
        - totalReferrals
        - creditsEarned
        - createdAt
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message