# lib/api-zod/src/generated/types/adminUserEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CraKa OSINT Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminUserEntry {
  referralCode: string;
  isPremium: boolean;
  premiumPlan?: string | null;
  totalReferrals: number;
  creditsEarned: number;
  createdAt: string;
}