# lib/api-zod/src/generated/types/grantPremiumRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CraKa OSINT Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface GrantPremiumRequest {
  referralCode: string;
  plan: string;
}