# lib/api-zod/src/generated/types/historyResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CraKa OSINT Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { HistoryEntry } from "./historyEntry";

export interface HistoryResponse {
  entries: HistoryEntry[];
  total: number;
  page: number;
  limit: number;
}