# lib/api-zod/src/generated/types/updateApiRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CraKa OSINT Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateApiRequest {
  name?: string;
  url?: string;
  command?: string;
  example?: string;
  pattern?: string | null;
  category?: string;
  credits?: number;
  isActive?: boolean;
}